factor out <Command.Bar>
